+++ util/string_test.go
@@ -242,3 +242,5 @@ func TestTextToFloat32(t *testing.T) {
 		})
 	}
 }
+
+// TODO: Write TestRFC3339ToTime